add new function of genres
