+++ src/index.js
@@ -1,4 +1,5 @@
 import templateFunction from './templates/card.hbs';
+import fetchGenres from './js/fetchGenres';
 const axios = require('axios').default;
 const wraper = document.querySelector('.div');
 import arrayGanre from './js/getGenre';
@@ -17,31 +18,20 @@ async function getUser() {
     const card = response.data.results[1]
     console.log(card)
     const { poster_path, title, release_date, vote_average, genre_ids } = card;
-    let genre = []
-    if (genre_ids.length <= 2) { genre = [...genre_ids] }
-    else {
-      genre = [...genre_ids].slice(0, 2);
-    }
-    newGenre = []
-    for (item of arrayGanre) { 
-      genre.map((genr) => {
-        if (genr === item.id) { newGenre.push(item.name) }
-      });
-    }
-    newGenre = newGenre.join(', ')
+    fetchGenres(genre_ids)
     return { poster_path, title, release_date, vote_average, newGenre }
   } catch (error) {
     console.error(error);
   }
 }
 
-async function makeMarkup() { 
+async function makeMarkup() {
   try {
     const data = await getUser();
     wraper.innerHTML = templateFunction(data);
 
   }
-  catch { 
+  catch {
     console.error(error);
   }
 }
